refactor rss_reader, complete iteration 2
