--- rss_reader/time.py
@@ -1,1 +1,0 @@
-import time
